move ext schema to main subgraphs
