--- subgraphs/aave-v3/protocols/aave-v3-extended/config/networks/polygon/polygon.json
@@ -1,11 +1,0 @@
-{
-  "network": "matic",
-  "PoolAddressesProviderRegistry": {
-    "address": "0x770ef9f4fe897e59daCc474EF11238303F9552b6",
-    "startBlock": 25824309
-  },
-  "RewardsController": {
-    "address": "0x929EC64c34a17401F460460D4B9390518E5B473e",
-    "startBlock": 25970800
-  }
-}
