+++ subgraphs/aave-v3/protocols/aave-v3/src/mappings/Pool.ts
@@ -6,9 +6,13 @@ import {
   Pool,
   Repay,
   ReserveDataUpdated,
+  ReserveUsedAsCollateralDisabled,
+  ReserveUsedAsCollateralEnabled,
   Supply,
   Withdraw,
 } from "../../../../generated/templates/Pool/Pool";
+import { BIGINT_ZERO } from "../../../../src/utils/constants";
+import { rayDiv } from "../../../../src/utils/numbers";
 import {
   createBorrow,
   createDeposit,
@@ -17,15 +21,14 @@ import {
   createWithdraw,
 } from "../entities/event";
 import { updateMarketRates } from "../entities/market";
+import { setUserLenderPositionIsCollateral } from "../entities/position";
 import { updateReserveAccruedToTreasury } from "../entities/reserve";
-import { BIGINT_ZERO } from "../../../../src/utils/constants";
-import { rayDiv } from "../../../../src/utils/numbers";
 
 export function handleBorrow(event: Borrow): void {
   createBorrow(
     event,
     event.params.reserve,
-    event.params.user,
+    event.params.onBehalfOf,
     event.params.amount
   );
 }
@@ -60,7 +63,7 @@ export function handleRepay(event: Repay): void {
   createRepay(
     event,
     event.params.reserve,
-    event.params.repayer,
+    event.params.user,
     event.params.amount
   );
 }
@@ -86,7 +89,7 @@ export function handleSupply(event: Supply): void {
   createDeposit(
     event,
     event.params.reserve,
-    event.params.user,
+    event.params.onBehalfOf,
     event.params.amount
   );
 }
@@ -95,7 +98,27 @@ export function handleWithdraw(event: Withdraw): void {
   createWithdraw(
     event,
     event.params.reserve,
-    event.params.to,
+    event.params.user,
     event.params.amount
+  );
+}
+
+export function handleReserveUsedAsCollateralEnabled(
+  event: ReserveUsedAsCollateralEnabled
+): void {
+  setUserLenderPositionIsCollateral(
+    event.params.user,
+    event.params.reserve,
+    true
+  );
+}
+
+export function handleReserveUsedAsCollateralDisabled(
+  event: ReserveUsedAsCollateralDisabled
+): void {
+  setUserLenderPositionIsCollateral(
+    event.params.user,
+    event.params.reserve,
+    false
   );
 }
