+++ subgraphs/aave-v3/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "@graphprotocol/graph-ts/types/tsconfig.base.json",
-  "include": ["src", "protocols/aave-v3/src", "protocols/aave-v3-extended/src"]
+  "include": ["src", "protocols/aave-v3/src"]
 }
